Revert "move types to internal-types"
This reverts commit 0eea74522e32ceaa656d0e6edfaa44d897bb0c65.
