+++ plugins/node/opentelemetry-instrumentation-koa/src/instrumentation.ts
@@ -23,11 +23,16 @@ import {
 } from '@opentelemetry/instrumentation';
 
 import type * as koa from 'koa';
-import { KoaInstrumentationConfig } from './types';
+import {
+  KoaMiddleware,
+  KoaContext,
+  KoaLayerType,
+  KoaInstrumentationConfig,
+} from './types';
 import { VERSION } from './version';
 import { getMiddlewareMetadata, isLayerIgnored } from './utils';
 import { getRPCMetadata, RPCType } from '@opentelemetry/core';
-import { KoaMiddleware, KoaLayerType, KoaContext, kLayerPatched, KoaPatchedMiddleware } from './internal-types';
+import { kLayerPatched, KoaPatchedMiddleware } from './internal-types';
 
 /** Koa instrumentation for OpenTelemetry */
 export class KoaInstrumentation extends InstrumentationBase<typeof koa> {
